+++ packages/docz-core/src/commands/build.ts
@@ -12,20 +12,18 @@ import { Config } from './args'
 export const build = async (args: Config) => {
   const env = envDotProp.get('node.env')
   const config = loadConfig(args)
+  const entries = new Entries(config)
+
   const bundler = webpack(config, env)
   const run = Plugin.runPluginsMethod(config.plugins)
   const dataServer = new DataServer()
 
+  dataServer.register([states.config(config), states.entries(entries, config)])
+
   try {
-    dataServer.register([
-      states.entries(config),
-      states.config(config),
-      states.metadata(config)
-    ])
-  
     await Entries.writeApp(config)
     await dataServer.init()
-return
+    return
     await run('onPreBuild', config)
     await bundler.build(await bundler.getConfig(env))
     await run('onPostBuild', config)