+++ packages/docz-core/src/config/mdx.ts
@@ -1,29 +1,14 @@
 import matter from 'remark-frontmatter'
 import slug from 'rehype-slug'
-import headings from 'rehype-autolink-headings'
 import remarkDocz from 'remark-docz'
 import rehypeDocz from 'rehype-docz'
+
+import * as paths from './paths'
 
 export const config = {
   type: 'yaml',
   marker: '-',
-  properties: {
-    'aria-hidden': true,
-  },
-  content: {
-    type: 'element',
-    tagName: 'span',
-    properties: {
-      className: ['icon-link'],
-    },
-    children: [
-      {
-        type: 'text',
-        value: '#',
-      },
-    ],
-  },
 }
 
 export const remarkPlugins = [matter, remarkDocz]
-export const rehypePlugins = [rehypeDocz, slug, headings]
+export const rehypePlugins = [rehypeDocz(paths.root), slug]