+++ packages/docz-core/src/states/config.ts
@@ -1,11 +1,10 @@
 import * as fs from 'fs-extra'
 import { load, finds } from 'load-cfg'
 import chokidar from 'chokidar'
-import equal from 'fast-deep-equal'
 import get from 'lodash.get'
 
 import { Params, State } from '../DataServer'
-import { Config, ThemeConfig } from '../commands/args'
+import { Config, Menu, ThemeConfig } from '../commands/args'
 import { getRepoUrl } from '../utils/repo-info'
 import * as paths from '../config/paths'
 
@@ -14,8 +13,10 @@ interface Payload {
   description: string
   ordering: string
   themeConfig: ThemeConfig
+  menu: Menu[]
   version: string | null
   repository: string | null
+  native: boolean
 }
 
 const getInitialConfig = (config: Config): Payload => {
@@ -26,34 +27,35 @@ const getInitialConfig = (config: Config): Payload => {
     title: config.title,
     description: config.description,
     themeConfig: config.themeConfig,
+    menu: config.menu,
     ordering: config.ordering,
     version: get(pkg, 'version'),
     repository: repoUrl,
+    native: config.native,
   }
 }
 
-const updateConfig = (config: Config) => async (p: Params) => {
-  const old = p.state.config
-  const newConfig = load('docz', getInitialConfig(config), true)
+const updateConfig = (config: Config) => async ({ setState }: Params) =>
+  setState('config', load('docz', getInitialConfig(config), true, false))
 
-  if (newConfig && !equal(old, newConfig)) {
-    p.setState('config', newConfig)
+export const state = (config: Config): State => {
+  const watcher = chokidar.watch(finds('docz'), {
+    cwd: paths.root,
+    ignored: /(^|[\/\\])\../,
+    persistent: true,
+  })
+
+  return {
+    init: updateConfig(config),
+    update: async params => {
+      const update = updateConfig(config)
+      const fn = async () => update(params)
+
+      watcher.on('add', fn)
+      watcher.on('change', fn)
+      watcher.on('unlink', fn)
+
+      return () => watcher.close()
+    },
   }
 }
-
-export const state = (config: Config): State => ({
-  init: updateConfig(config),
-  update: async params => {
-    const update = updateConfig(config)
-    const watcher = chokidar.watch(finds('docz'), {
-      cwd: paths.root,
-      persistent: true,
-    })
-
-    watcher.on('add', async () => update(params))
-    watcher.on('change', async () => update(params))
-    watcher.on('unlink', async () => update(params))
-
-    return () => watcher.close()
-  },
-})