+++ packages/docz-theme-default/src/components/ui/Editor/index.tsx
@@ -26,9 +26,15 @@ import 'codemirror/addon/fold/xml-fold'
 const getLanguage = (children: any) => {
   const defaultLanguage = 'jsx'
   if (typeof children === 'string') return defaultLanguage
+
   const language = get(children, 'props.props.className') || defaultLanguage
   const result = language.replace('language-', '')
-  return result === 'js' || result === 'javascript' ? 'jsx' : result
+
+  if (result === 'js' || result === 'javascript') return 'jsx'
+  if (result === 'ts' || result === 'tsx' || result === 'typescript') {
+    return 'text/typescript'
+  }
+  return result
 }
 
 const getChildren = (children: any) =>
@@ -128,7 +134,7 @@ export const ClipboardAction: SFC<ClipboardActionProps> = ({
 }) => (
   <ActionButton
     {...props}
-    as={ButtonLink}
+    as={ButtonLink as any}
     title="Copy to clipboard"
     onClick={() => copy(content)}
     swap={<Check width={17} />}