+++ packages/docz-theme-default/webpack.config.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const webpack = require('webpack')
-const UglifyJs = require('uglifyjs-webpack-plugin')
+const TerserPlugin = require('terser-webpack-plugin')
 const FileManagerPlugin = require('filemanager-webpack-plugin')
 const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer')
 
@@ -18,6 +18,7 @@ const externalList = [
   'codemirror/addon/edit/matchbrackets',
   'codemirror/addon/edit/closetag',
   'codemirror/addon/fold/xml-fold',
+  'normalize.css',
   'react-perfect-scrollbar',
   'polished/lib/color/rgba',
   'polished/lib/color/lighten',
@@ -36,19 +37,22 @@ const externalList = [
   'react-feather/dist/icons/maximize',
   'react-feather/dist/icons/minimize',
   'react-feather/dist/icons/refresh-cw',
+  'react-feather/dist/icons/hash',
   'react-sizes',
 ]
 
-const deps = Object.keys(pkg.dependencies)
-const externals = Object.keys(pkg.dependencies)
-  .concat(externalList)
-  .concat(deps.filter(dep => dep.startsWith('react-feather')))
+const internals = [
+  'normalize.css',
+  'codemirror/lib/codemirror.css'
+]
 
-const uglify = new UglifyJs({
-  parallel: true,
-  cache: true,
-  sourceMap: true,
-  uglifyOptions: {
+const deps = Object.keys(pkg.dependencies)
+const externals = deps
+  .concat(externalList)
+  .filter(dep => internals.indexOf(dep) === -1)
+
+const minify = new TerserPlugin({
+  terserOptions: {
     parse: {
       ecma: 8,
     },
@@ -66,6 +70,9 @@ const uglify = new UglifyJs({
       ascii_only: true,
     },
   },
+  parallel: true,
+  cache: true,
+  sourceMap: true,
 })
 
 const plugins = [
@@ -106,6 +113,10 @@ module.exports = {
           },
         ],
       },
+      {
+        test: /\.css$/,
+        use: [ 'style-loader', 'css-loader' ]
+      }
     ],
   },
   resolve: {
@@ -121,7 +132,7 @@ module.exports = {
     namedModules: true,
     ...(IS_PROD && {
       minimize: true,
-      minimizer: [uglify],
+      minimizer: [minify],
     }),
   },
   performance: {