+++ packages/docz/src/components/Docs.tsx
@@ -1,32 +1,18 @@
 import * as React from 'react'
-import { Children } from 'react'
 import sort from 'array-sort'
 
 import { state, Entry, EntryMap, Config } from '../state'
 import { entriesSelector } from './DocPreview'
 import { configSelector } from './ThemeConfig'
-
-export const isFn = (value: any): boolean => typeof value === 'function'
-
-const sortBy = (a: any, b: any, reverse?: boolean) => {
-  if (a < b) return reverse ? 1 : -1
-  if (a > b) return reverse ? -1 : 1
-  return 0
-}
-
-const menuFromEntries = (entries: Entry[]) =>
-  Array.from(
-    new Set(
-      entries.reduce(
-        (arr: string[], entry: Entry): string[] =>
-          entry.menu ? arr.concat([entry.menu]) : arr,
-        []
-      )
-    )
-  )
+import { compare, flatArrFromObject, isFn } from '../utils/helpers'
 
 export interface DocsRenderProps {
   docs: Entry[]
+  /**
+   * @deprecated Since the new <Menu> component now support
+   * ordering by set menu property on config file
+   * will be deleted in v1.0
+   */
   menus: string[]
 }
 
@@ -48,22 +34,21 @@ export const Docs: React.SFC<DocsProps> = ({ children }) => {
         }
 
         const arr = Object.values(entries)
-        const menusArr = menuFromEntries(arr)
-        const menus = sort(menusArr, (a: Entry, b: Entry) => sortBy(a, b))
-        const descending = config.ordering === 'descending'
 
+        /** TODO: remove all order and  logic from here in a breaking change */
+        const menusArr = flatArrFromObject<Entry>(arr, 'menu')
+        const menus = sort(menusArr, (a: string, b: string) => compare(a, b))
+        const descending = config.ordering === 'descending'
         const docs: Entry[] = sort(
           arr,
-          (a: Entry, b: Entry) => sortBy(a.order, b.order, descending),
-          (a: Entry, b: Entry) => sortBy(a.name, b.name)
+          (a: Entry, b: Entry) => compare(a.order, b.order, descending),
+          (a: Entry, b: Entry) => compare(a.name, b.name)
         )
 
-        return Children.only(
-          children({
-            menus,
-            docs,
-          })
-        )
+        return children({
+          menus,
+          docs,
+        })
       }}
     </state.Consumer>
   )