+++ packages/rehype-docz/src/index.ts
@@ -1,66 +1,54 @@
+import * as path from 'path'
 import is from 'unist-util-is'
-import nodeToString from 'hast-util-to-string'
-import strip from 'strip-indent'
 import flatten from 'lodash.flatten'
-
-import { codeFromNode } from './code-from-node'
-import { importsFromEntry } from './imports-from-entry'
-import { format } from './format'
-
-const componentName = (value: any) => {
-  const match = value.match(/^\<\\?(\w+)/)
-  return match && match[1]
-}
+import nodeToString from 'hast-util-to-string'
+import { jsx, imports as imps } from 'docz-utils'
+import { format } from 'docz-utils/lib/format'
+import { getSandboxImportInfo } from 'docz-utils/lib/codesandbox'
 
 const isPlayground = (name: string) => name === 'Playground'
 
-const isOpenTag = (p: any) =>
-  p.isJSXOpeningElement() && isPlayground(p.node.name.name)
-
-const isClosetag = (p: any) =>
-  p.isJSXClosingElement() && isPlayground(p.node.name.name)
-
-const removePlayground = (code: string) => {
-  const open = codeFromNode(isOpenTag)
-  const close = codeFromNode(isClosetag)
-
-  return code.replace(open(code), '').replace(close(code), '')
-}
-
-const addCodeProp = (imports: string[]) => async (node: any, idx: number) => {
-  const name = componentName(node.value)
+const addPropsOnPlayground = async (
+  node: any,
+  idx: number,
+  scopes: string[],
+  imports: string[],
+  cwd: string
+) => {
+  const name = jsx.componentName(node.value)
   const tagOpen = new RegExp(`^\\<${name}`)
 
   if (isPlayground(name)) {
     const formatted = await format(nodeToString(node))
     const code = formatted.slice(1, Infinity)
-    const scope = `{props,${imports.join(',')}}`
-    const child = strip(removePlayground(code))
-      .trim()
-      .replace(/'/g, `\\'`)
-      .replace(/`/g, '\\`')
+    const scope = `{props,${scopes.join(',')}}`
+    const child = jsx.sanitizeCode(jsx.removeTags(code))
+    const codesandBoxInfo = await getSandboxImportInfo(child, imports, cwd)
 
     node.value = node.value.replace(
       tagOpen,
-      `<${name} __position={${idx}} __code={\`${child}\`} __scope={${scope}}`
+      `<${name} __position={${idx}} __codesandbox={\`${codesandBoxInfo}\`} __code={\`${child}\`} __scope={${scope}}`
     )
   }
 }
 
-const getScope = (node: any, idx: number) => {
-  return importsFromEntry(node)
+const addComponentsProps = (
+  scopes: string[],
+  imports: string[],
+  cwd: string
+) => async (node: any, idx: number) => {
+  await addPropsOnPlayground(node, idx, scopes, imports, cwd)
 }
 
-export default () => (tree: any) => {
-  const imports: string[] = flatten(
-    tree.children.filter((node: any) => is('import', node)).map(getScope)
-  )
+export default (root: string) => () => (tree: any, fileInfo: any) => {
+  const importNodes = tree.children.filter((node: any) => is('import', node))
+  const imports: string[] = flatten(importNodes.map(imps.getFullImports))
+  const scopes: string[] = flatten(importNodes.map(imps.getImportsVariables))
+  const fileCwd = path.relative(root, path.dirname(fileInfo.history[0]))
 
-  const nodes = [
-    tree.children
-      .filter((node: any) => is('jsx', node))
-      .map(addCodeProp(imports)),
-  ]
+  const nodes = tree.children
+    .filter((node: any) => is('jsx', node))
+    .map(addComponentsProps(scopes, imports, fileCwd))
 
   return Promise.all(nodes).then(() => tree)
 }